+++ mitra_api/src/mastodon_api/subscriptions/views.rs
@@ -213,8 +213,7 @@ async fn register_subscription_option(
             let price: NonZeroU64 = price.try_into()
                 .map_err(|_| ValidationError("price must be greater than 0"))?;
             validate_subscription_price(price)?;
-            validate_monero_address(&payout_address, &chain_id)
-                .map_err(|_| ValidationError("invalid monero address"))?;
+            validate_monero_address(&payout_address, &chain_id)?;
             let payment_method_data = PaymentMethodData {
                 owner_id: current_user.id,
                 payment_type: PaymentType::Monero,
@@ -237,8 +236,7 @@ async fn register_subscription_option(
             let price: NonZeroU64 = price.try_into()
                 .map_err(|_| ValidationError("price must be greater than 0"))?;
             validate_subscription_price(price)?;
-            validate_monero_standard_address(&payout_address, &chain_id)
-                .map_err(|_| ValidationError("invalid monero standard address"))?;
+            validate_monero_standard_address(&payout_address, &chain_id)?;
             let payout_address = parse_monero_address(&payout_address)
                 .map_err(|_| ValidationError("invalid monero address"))?;
             let view_key = parse_monero_view_key(&view_key)